TP4/3-CloudRunFunction/requirements.txt
functions-framework==3.*
tensorflow
librosa
google-cloud-storage
